+++ oraclepaas/resource_application_container.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"log"
 	"strconv"
+	"strings"
 	"time"
 
 	"github.com/hashicorp/go-oracle-terraform/application"
@@ -270,6 +271,25 @@ func resourceOraclePAASApplicationContainer() *schema.Resource {
 				ForceNew:  true,
 				Sensitive: true,
 			},
+			"availability_domain": {
+				Type:     schema.TypeList,
+				Optional: true,
+				ForceNew: true,
+				Elem:     &schema.Schema{Type: schema.TypeString},
+			},
+			"load_balancer_subnets": {
+				Type:     schema.TypeList,
+				MaxItems: 2,
+				MinItems: 2,
+				Optional: true,
+				ForceNew: true,
+				Elem:     &schema.Schema{Type: schema.TypeString},
+			},
+			"region": {
+				Type:     schema.TypeString,
+				Optional: true,
+				ForceNew: true,
+			},
 			"app_url": {
 				Type:     schema.TypeString,
 				Computed: true,
@@ -329,6 +349,18 @@ func resourceOraclePAASApplicationContainerCreate(d *schema.ResourceData, meta i
 
 	if v, ok := d.GetOk("git_password"); ok {
 		additionalFields.GitPassword = v.(string)
+	}
+
+	if _, ok := d.GetOk("availability_domain"); ok {
+		additionalFields.AvailabilityDomain = strings.Join(getStringList(d, "availability_domain"), ",")
+	}
+
+	if _, ok := d.GetOk("load_balancer_subnets"); ok {
+		additionalFields.LoadBalancerSubnets = strings.Join(getStringList(d, "load_balancer_subnets"), ",")
+	}
+
+	if v, ok := d.GetOk("region"); ok {
+		additionalFields.Region = v.(string)
 	}
 
 	if v, ok := d.GetOk("tags"); ok {